Add support for Linux/Wayland with wl-paste
This modifies linux.sh to support Wayland systems that do not use the
xclip command.

Logic is added to attempt to determine the type of display in use
(X11/Xorg or Wayland), and based on that, try the common clipboard CLI
commands for the given display system.

Failing to determine the display system for any reason will fall back to
the original behavior that assumes xclip is present.

This change also resulted in refactoring the script a bit since it got a
bit more complicated, any style changes are in-line with the Google
shell style guide (https://google.github.io/styleguide/shellguide.html)

A minor change to extension.ts was also made to support the error case
of wl-paste note being present on Wayland systems.

This change should resolve https://github.com/dendronhq/dendron/issues/1095